update dependencies
This is a large commit to bring most dependencies up to date.

phpunit runs with no errors, but some warnings (deprecations that need
to be addressed but not immediately)
apigen has been replaced by phpdoc because it seems unmaintained.

The travis build no longer builds older versions of PHP (<7.3). This
is a breaking release, and should likely be bumped to semver 2.0.0,
and this is reflected in the Akamai_Open_EdgeGrid_Client::VERSION
constant.